+++ packages/map-styles/generate-styles.ts
@@ -1,4 +1,4 @@
-import  { createLineSymbolLayers } from "./src" ;
+import { createLineSymbolLayers } from "./src";
 
 let output = createLineSymbolLayers();
 console.log(output);