+++ packages/map-styles/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2026-01-26
+
+- Removed import of `mapbox-gl/dist/mapbox-gl.css`, no styles bundled with
+  library.
+- Changed bundling of library to use `@macrostrat/web-components-bundler`.
+- Updated peer dependencies to React 18.
+
 ## [1.2.4] - 2025-11-28
 
 - Fix a NaN comparison bug in resolveFGDCImage